Simplify CollisionAlgorithmProvider interface, and repair all currently implemented Shapes.

+++ ferox-physics/src/main/java/com/ferox/physics/collision/Shape.java
@@ -66,5 +66,5 @@ public interface Shape {
     /**
      * @return Return the current margin for this shape, defaults to .05
      */
-    public float getMargin();
+    public double getMargin();
 }
